Moved graph renderers to their own layers
