+++ src/app/modules/fsa2/FSAModule.js
@@ -31,18 +31,10 @@ import OverviewPanel from './components/panels/overview/OverviewPanel.js';
 import TestingPanel from './components/panels/testing/TestingPanel.js';
 import AnalysisPanel from './components/panels/analysis/AnalysisPanel.js';
 
-import FSANodeRenderer from './graph/renderer/FSANodeRenderer.js';
-import FSANodeInputHandler from './graph/inputhandler/FSANodeInputHandler.js';
-import FSAInitialMarkerLayer from './graph/renderer/FSAInitialMarkerLayer.js';
-import FSALabelEditorRenderer from './renderer/FSALabelEditorRenderer.js';
+import FSAGraphLayer from './components/FSAGraphLayer.js';
+import FSAGraphOverlayLayer from './components/FSAGraphOverlayLayer.js';
 
 import GraphView from 'graph2/components/GraphView.js';
-import GraphNodeLayer from 'graph2/components/layers/GraphNodeLayer.js';
-import GraphEdgeLayer from 'graph2/components/layers/GraphEdgeLayer.js';
-import SelectionBoxLayer from 'graph2/components/layers/SelectionBoxLayer.js';
-import ViewportLayer from 'graph2/components/layers/ViewportLayer.js';
-import ViewportNavigationLayer from 'graph2/components/layers/ViewportNavigationLayer.js';
-import LabelEditorView from 'graph2/components/views/LabelEditorView.js';
 
 const MODULE_NAME = 'fsa2';
 const MODULE_VERSION = '0.0.1';
@@ -56,67 +48,17 @@ class FSAModule
         this._graph = new FSAGraph();
         this._graphController = new FSAGraphController(app, this._graph, null);
         this._graphViewComponent = React.createRef();
-
-        const graph = this._graph;
+        
         const graphController = this._graphController;
-        const labelFormatter = graphController.getLabelFormatter();
 
         app.getRenderManager()
             .addRenderer(RENDER_LAYER_WORKSPACE, props => (
                 <GraphView
                     ref={this._graphViewComponent}
-                    renderGraph={graphView => (
-                        <React.Fragment>
-                            <FSAInitialMarkerLayer
-                                inputController={graphView.getInputController()}
-                                graphController={graphController}
-                                inputContext={graphView.getInputContext()}
-                                inputPriority={-1} />
-                            <GraphNodeLayer
-                                nodes={graph.getNodes()}
-                                inputController={graphView.getInputController()}
-                                graphController={graphController}
-                                nodeRenderer={FSANodeRenderer}
-                                inputContext={graphView.getInputContext()}
-                                inputPriority={-1} />
-                            <GraphEdgeLayer
-                                ref={ref =>
-                                {
-                                    if (!ref) return;
-                                    ref.getGraphEdgeInputHandler()
-                                        .setShouldDeleteEdgeWithEmptyLabel(true)
-                                        .setShouldDeleteEdgePlaceholder(true);
-                                }}
-                                edges={graph.getEdges()}
-                                inputController={graphView.getInputController()}
-                                graphController={graphController}
-                                inputContext={graphView.getInputContext()}
-                                inputPriority={-1} />
-                            <SelectionBoxLayer
-                                inputController={graphView.getInputController()}
-                                graphController={graphController}
-                                inputContext={graphView.getInputContext()}
-                                inputPriority={-1} />
-                        </React.Fragment>
-                    )}
-                    renderOverlay={graphView => (
-                        <React.Fragment>
-                            <ViewportLayer
-                                graphController={graphController}
-                                inputController={graphView.getInputController()}
-                                viewport={graphView.getViewportComponent()}>
-                                <ViewportNavigationLayer
-                                    style={{ right: 0 }}
-                                    viewportAdapter={graphView.getViewportComponent().getViewportAdapter()} />
-                            </ViewportLayer>
-                            <LabelEditorView ref={ref => graphController.setLabelEditor(ref)}
-                                labelFormatter={labelFormatter}
-                                viewport={graphView.getViewportComponent()}
-                                saveOnExit={true}>
-                                <FSALabelEditorRenderer graphController={graphController} currentModule={this} />
-                            </LabelEditorView>
-                        </React.Fragment>
-                    )}>
+                    renderGraph={graphView =>
+                        <FSAGraphLayer graphView={graphView} graphController={graphController} />}
+                    renderOverlay={graphView =>
+                        <FSAGraphOverlayLayer graphView={graphView} graphController={graphController} module={this} />}>
                 </GraphView>
             ));
 
@@ -133,15 +75,9 @@ class FSAModule
     /** @override */
     initialize(app)
     {
-        const graphView = this._graphViewComponent.current;
-        const inputController = graphView.getInputController();
-        const inputContext = graphView.getInputContext();
-        const graphController = this._graphController;
-        inputContext.addInputHandler(new FSANodeInputHandler(inputController, graphController));
-
         registerNotifications(app.getNotificationManager());
 
-        //TODO: These should have a pre/post handlers...
+        // TODO: These should have a pre/post handlers...
         app.getExportManager()
             .registerExporter(new FSAExporter(FSAGraphParser.JSON), 'session')
             .registerExporter(new FSAJFFExporter(FSAGraphParser.XML), 'jflap');
