+++ src/app/modules/fsa2/graph/widgets/FSALabelEditorRenderer.css
